Fix hardlinking of files to use link instead of copytree which doesn't work

+++ scripts/sweeper.py
@@ -156,7 +156,10 @@ class Sweeper:
             copytree(self._torrent.root, final_target)
         elif self._action == "hard":
             logger.info(f"HARDLINKING '{self._torrent.root}' to '{final_target}'")
-            copytree(self._torrent.root, final_target, copy_function=link)
+            if self._torrent.root.is_file:
+                link(self._torrent.root, final_target)
+            else:
+                copytree(self._torrent.root, final_target, copy_function=link)
         elif self._action == "test":
             logger.info(f"NOOP '{self._torrent.root}' to '{final_target}'")
         else:
